# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/money/customer/orders/{orderId}/refunds" : {
      "get" : {
        "tags" : [ "customer-money-controller" ],
        "operationId" : "getOrderRefunds",
        "parameters" : [ {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/RefundView"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/money/customer/orders/{orderId}/receipt" : {
      "get" : {
        "tags" : [ "customer-money-controller" ],
        "operationId" : "getReceipt",
        "parameters" : [ {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CustomerReceipt"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/internal/admin/orders/{orderId}/money" : {
      "get" : {
        "tags" : [ "admin-money-controller" ],
        "operationId" : "getOrderMoney",
        "parameters" : [ {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminOrderMoney"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "RefundView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "orderId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "amount" : {
            "type" : "number"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "REQUESTED", "PROCESSING", "COMPLETED", "FAILED", "CANCELLED" ]
          },
          "destination" : {
            "type" : "string",
            "enum" : [ "ORIGINAL_METHOD", "STORE_CREDIT", "NONE" ]
          },
          "method" : {
            "type" : "string",
            "enum" : [ "CARD", "UPI", "WALLET", "COD" ]
          },
          "reasonCode" : {
            "type" : "string"
          },
          "requestedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "completedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "expectedBy" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "CustomerReceipt" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ReceiptItem"
            }
          },
          "itemTotal" : {
            "type" : "number"
          },
          "deliveryFee" : {
            "type" : "number"
          },
          "platformFee" : {
            "type" : "number"
          },
          "sgst" : {
            "type" : "number"
          },
          "cgst" : {
            "type" : "number"
          },
          "total" : {
            "type" : "number"
          },
          "paymentMethod" : {
            "type" : "string"
          },
          "paidAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "refunds" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RefundView"
            }
          },
          "storeCreditUsed" : {
            "type" : "number"
          }
        }
      },
      "ReceiptItem" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "quantity" : {
            "type" : "integer",
            "format" : "int32"
          },
          "price" : {
            "type" : "number"
          }
        }
      },
...
      "AdminOrderMoney" : {
        "type" : "object",
        "properties" : {
          "orderId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "totalAmount" : {
            "type" : "number"
          },
          "foodCost" : {
            "type" : "number"
          },
          "deliveryFee" : {
            "type" : "number"
          },
          "customerPlatformFee" : {
            "type" : "number"
          },
          "restaurantPayout" : {
            "type" : "number"
          },
          "restaurantPlatformFee" : {
            "type" : "number"
          },
          "restaurantDeliveryContribution" : {
            "type" : "number"
          },
          "driverGrossPayout" : {
            "type" : "number"
          },
          "driverTaxes" : {
            "type" : "number"
          },
          "driverNetPayout" : {
            "type" : "number"
          },
          "platformBonus" : {
            "type" : "number"
          },
          "sgst" : {
            "type" : "number"
          },
          "cgst" : {
            "type" : "number"
          },
          "ledgerLines" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/LedgerStatementLineDto"
            }
          }
        }
      },
      "LedgerStatementLineDto" : {
        "type" : "object",
        "properties" : {
          "transactionId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "referenceId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "category" : {
            "type" : "string",
            "enum" : [ "DELIVERY_FEE", "PLATFORM_FIXED_FEE", "PLATFORM_BONUS", "FOOD_COST", "SGST", "CGST", "REFUND", "ORDER_TOTAL", "AD_IMPRESSION", "AD_CLICK", "AD_CONVERSION", "AD_WALLET_TOPUP", "CLAWBACK", "PAYOUT_TRANSFER", "CASH_COLLECTED", "CASH_REMITTED", "STORE_CREDIT" ]
          },
          "amount" : {
            "type" : "number"
          },
          "direction" : {
            "type" : "string",
            "enum" : [ "CREDIT", "DEBIT" ]
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "description" : {
            "type" : "string"
          },
          "payoutId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "payoutStatus" : {
            "type" : "string"
          },
          "settled" : {
            "type" : "boolean"
          }
        }
      },
...
    }
  }
}